Fixed quota comment

+++ yt/server/chunk_holder/config.h
@@ -48,10 +48,10 @@ struct TLocationConfig
         Register("quota", Quota).Default(TNullable<i64>());
         Register("low_watermark", LowWatermark)
             .GreaterThan(0)
-            .Default(20L * 1024 * 1024 * 1024); // 1 G
+            .Default((i64) 20 * 1024 * 1024 * 1024); // 20 Gb
         Register("high_watermark", HighWatermark)
             .GreaterThan(0)
-            .Default(10L * 1024 * 1024 * 1024); // 100 Mb
+            .Default((i64) 10 * 1024 * 1024 * 1024); // 10 Gb
     }
 
     virtual void DoValidate() const
